Test Catalog: removed unneeded query sample code from main codebase
